requirements.txt
aiogram==3.13.1
openai==1.82.0
openai-whisper
pydantic-settings==2.7.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
